# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "库里的演示账号：`neo` / `matrix`、`trinity` / `zion`。用户名去空格并转小写，**密码不做任何处理**（首尾空格是密码的一部分）。凭证不对时统一返回 401 与同一句文案 —— 不区分\"用户不存在\"和\"密码错\"，避免用户名枚举。",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证：缺少凭证、`Authorization` 不是 Bearer、凭证无效或已过期（响应头带 `WWW-Authenticate`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Unauthorized",
                    "message": "Missing or malformed credentials",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "登录并签发 JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "description": "把 token 里的声明原样回显：`sub` + `username`。不含角色 / 权限 —— 本模块只回答\"你是谁\"，不回答\"你能不能\"。",
        "operationId": "AuthController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "token 有效",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseEnvelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProfileDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参校验失败（`errors[]` 里是字段级明细）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Bad Request",
                    "message": "Request validation failed",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f",
                    "errors": [
                      {
                        "field": "name",
                        "location": "body",
                        "message": "name must be longer than or equal to 2 characters"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证：缺少凭证、`Authorization` 不是 Bearer、凭证无效或已过期（响应头带 `WWW-Authenticate`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Unauthorized",
                    "message": "Missing or malformed credentials",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源或路由不存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Not Found",
                    "message": "user 999999 not found",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁（响应头带 `Retry-After` 与 `X-RateLimit-*`）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Too Many Requests",
                    "message": "Too many requests, please try again later",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          },
          "500": {
            "description": "未捕获的内部错误（细节只进日志，不进响应）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "success": false,
                    "error": "Internal Server Error",
                    "message": "Internal server error",
                    "traceId": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
                  }
                }
              }
            }
          }
        },
        "summary": "当前身份（需要 Bearer token）",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "nestjs-start API",
    "description": "`nestjs-start` 的接口文档。\n\n**响应契约**：\n- 成功 `{ success: true, data, meta? }`\n- 失败 `{ success: false, error, message, traceId?, errors? }`\n\n`success` 是唯一判据；`code` 是机器判据（**不要 parse `message`**）；\n数字状态码只在 HTTP 状态行里；`traceId` 同时回写在 `x-request-id` 响应头。\n详见 `docs/validation.md` §9。",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名（去首尾空格、大小写不敏感）",
            "example": "neo"
          },
          "password": {
            "type": "string",
            "description": "密码（**不去空格**：首尾空格是密码的一部分）",
            "example": "matrix",
            "minLength": 1,
            "maxLength": 72
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌，后续用 `Authorization: Bearer <token>` 发送",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIn0.…"
          },
          "tokenType": {
            "type": "string",
            "description": "令牌类型，固定 `Bearer`",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "description": "有效期（秒），由 token 的 `exp - iat` 得出",
            "example": 3600
          }
        },
        "required": [
          "accessToken",
          "tokenType"
        ]
      },
      "ProfileDto": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "string",
            "description": "主体 id（JWT 的 `sub`）",
            "example": "1"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "neo"
          }
        },
        "required": [
          "sub",
          "username"
        ]
      },
      "ResponseEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "description": "handler 的返回值（分页时是这一页的数据数组）"
          },
          "meta": {
            "description": "仅分页成功时出现",
            "type": "object",
            "required": [
              "totalItems",
              "itemsPerPage",
              "currentPage"
            ],
            "properties": {
              "totalItems": {
                "type": "integer",
                "example": 3
              },
              "itemsPerPage": {
                "type": "integer",
                "example": 50
              },
              "currentPage": {
                "type": "integer",
                "example": 1
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "traceId": {
            "description": "请求 id：同时回写在 `x-request-id` 响应头、并写进服务端日志，用来把一次响应和日志对上。",
            "type": "string",
            "example": "3f1c9a4e-6b0a-4e2f-9a1d-6a1b1c2d3e4f"
          },
          "errors": {
            "description": "仅校验类失败才有（`field` 是权威定位，嵌套用点号）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "出问题的字段路径，嵌套时用点号",
            "example": "address.city"
          },
          "message": {
            "type": "string",
            "description": "人类可读说明（`message` 只给概述，具体规则看这里）",
            "example": "city must be longer than or equal to 2 characters"
          },
          "location": {
            "description": "这个字段路径来自哪里（`@Body()` / `@Query()` / `@Param()` 各不同）。同名不同源时靠它区分。",
            "type": "string",
            "enum": [
              "body",
              "query",
              "param"
            ],
            "example": "body"
          }
        }
      }
    }
  }
}